Partially concealable italics/bold/code
Delimiters must reside in the same line. Concealment is limited to:

- sequences of non-space characters of any kind
- groups of (keyword) characters, separated by spaces

Enable it by setting g:asciidoctor_syntax_concealed to 1.

The new groups are marked as 'contained', to avoid their inclusion in
non-paragraph blocks. A couple syntax groups had to be created for this
to work (asciidoctorParagraph and asciidoctorList).

I only tested with the documents I work with, not sure if there are side
effects, it's disabled by default anyway.